Fixed bug in GenerateSchedule.lhs where windowed projects were not getting their own unique pId.
Before, all windowed projects had the same pId == 0.  As a consequence of this, Pack would only schedule *one* chosen period for one of n windows in a single call, since it would think the project's time (the one project for all windows) was used up after that one chosen period.