Improved the performance of getPeriodObsEffFactors in Statistics.lhs by 10x !
Before, we were making sure we calculated the efficiencies for when the period was observed by setting the weather origin for each quarter in the period; this has tremendour overhead.  Instead, we can get the same result (checked by new unit test) simply setting the origin once for each period at the END of the period.