Fixed two bugs in plotFreqVsTimeWindowed ('simFreqTimeWindowed.jpg')
First, the two different data sets weren't using the same definition of 'day 0', so had to create the historicalTime'From function.
Second, when we started using the altered sessions from the simulations, rather then the original onces, that broke how this plot got the scheduled default periods - so now we're using the trace.